18th commit

+++ src/main/java/alex/bandcloud/repos/BandRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BandRepo extends CrudRepository<Band, Long> {
+public interface BandRepo extends CrudRepository<Band, Integer> {
     Band findFirstByName(String name);
 }
